updated
updated the code inside if(isset($_SESSION["index"])) so that it helps create a session variable needed to validate sessions through rabbit mq. I execute a select statement to grab the sessionID from my mysql table.

I created a code to allow api calls to go through rabbit mq. Whenever any of my webpages calls an api, i create a session called $_SESSION["apiName"] and assign it a string that dictates the purpose of the api so that when testRabbitMQServer.php is listening it will know when and what apis are being called. $_SESSION["apiName"] is sent from testRabbitMQClient.php to testRabbitMQServer.php