Updated so that a users disliked and like restaurants impact their suggested food and have a running history of food so that users cant get the same food twice
added $_SESSION["apiName"] so that we can run api through rabbitmq. 
added a shuffle code that will always shuffle the array of fast food restaurants i search from. 
After shuffling the array i make it go through if statements that will execute depending on what restaurants the users like and dislikes. 

if they dislike a restaurants, i use unset to remove that restaurant from the default array. 

If they like a restaurant, i put it in the first index of the array so it tries to suggest a food item from the restaurant first.

I updated the order of the for loop so now it loops through the default array i made, and then loops into the 20 google maps api restaurants. Previously i had it the other way around.

With this way, i am able to get more variety in the suggested fast food items because now we always shuffle the default array and modify the array based on the users likes and dislikes.

Added the sql code that retrieves a users latest food they have been suggested. I run a for loop that will exit the program if a user gets the same food suggestion twice in a row.

added the sql code that inserts the food that i suggest a user ,into a sql table called historyFood

added a if statement at the end of the program that basically will exit the program if my code was unable to suggest a food to a user.