Added a likes record page to website
This page allows users to see all the currently liked and disliked restaurants under their username.
There are two tables that are displayed. Under the liked restaurants theres is a hyperlink to resetLikes.php and under the disliked table there is a hyperlink to resetDislikes.php which allows users to reset their likes / dislikes for restaurants.